+++ model/Die.h
@@ -2,18 +2,18 @@
 #define DIE_H_FA033981
 
 typedef struct {
-	int num_faces;
-	int last_result;
+	size_t num_faces;
+	size_t last_result;
 } Die;
 
-Die * die_create(int faces);
+Die * die_create(size_t faces);
 
 void die_free(Die *d);
 
-int die_roll(Die *d);
+size_t die_roll(Die *d);
 
-int die_last_result(Die *d);
-int die_faces(Die *d);
+size_t die_last_result(Die *d);
+size_t die_faces(Die *d);
 
 
 #endif /* end of include guard: DIE_H_FA033981 */
